refactor: Jdbc 레포지토리 코드 제거 및 Jpa 레포지토리로 통합
- Jdbc 기반의 Reservation, ReservationTime, Theme, User 레포지토리 코드를 삭제하고 기존 Jpa 기반 레포지토리로 통합.
- 관련 테스트 코드 및 인프라 패키지 구조에서 불필요한 "Jpa" 네이밍 제거로 간결화 작업 수행.
